% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{catch}
\alias{catch}
\title{Catches errors from a given expression. Can be used in a pipe.}
\usage{
catch(...)
}
\arguments{
\item{...}{The expression to safely evaluate.}
}
\value{
The result of evaluating the expression or NULL in case of error.
}
\description{
Catches errors from a given expression. Can be used in a pipe.
}
\keyword{internal}
\keyword{utils}
