% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate.R
\name{check_known_packages}
\alias{check_known_packages}
\title{Seaches a list of known package hex links for a given package name}
\usage{
check_known_packages(
  pkg_name,
  known_packages = getOption("hexFinder.known_packages")
)
}
\arguments{
\item{pkg_name}{A single string with the name of a package to find
the hex for.}

\item{known_packages}{Named list with known hex locations for some packages.}
}
\value{
A URL that points to a image or NULL in no result was found.
}
\description{
Some packages have logos but are not tied to CRAN in any way. These are
  stored as special cases.
}
\keyword{internal}
\keyword{locator}
