% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate.R
\name{check_known_repos}
\alias{check_known_repos}
\title{Searches a list of known package hex curated lists for a given package name}
\usage{
check_known_repos(pkg_name, known_repos = getOption("hexFinder.known_repos"))
}
\arguments{
\item{pkg_name}{A single string with the name of a package to find
the hex for.}

\item{known_repos}{Named list with known curated repos that store logos
for some packages. quality on these is usually higher that normal.}
}
\value{
A URL that points to a image or NULL in no result was found.
}
\description{
Some packages have logos already in curated repositories. This avoids
  additional CRAN calls and returns high definition versions of the logos.
}
\keyword{internal}
\keyword{locator}
