% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{clean_repo_links}
\alias{clean_repo_links}
\title{Returns gitub repo links based on given URLs}
\usage{
clean_repo_links(...)
}
\arguments{
\item{...}{Any links we would like to parse to find a github URL from.}
}
\value{
A list of possible github repos
}
\description{
Returns gitub repo links based on given URLs
}
\keyword{internal}
\keyword{utils}
