% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{crop_image}
\alias{crop_image}
\title{Crops the empty space from a image}
\usage{
crop_image(path)
}
\arguments{
\item{path}{The path of the image to crop.}
}
\value{
No return, called for side effects.
}
\description{
Crops the empty space from a image
}
\keyword{images}
\keyword{internal}
