% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{download_logo}
\alias{download_logo}
\title{Download a logo image}
\usage{
download_logo(pkg_name, raw_url, output, crop = TRUE)
}
\arguments{
\item{pkg_name}{The name of the file to save.}

\item{raw_url}{The image URL.}

\item{output}{The Folder path where to store the download file.}

\item{crop}{Should empty space be removed from the final image file.}
}
\value{
The path where the downloaded image was stored.
}
\description{
Downloads a given URL and stores it using the package name into
the specified output folder
}
\keyword{images}
\keyword{internal}
