% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/images.R
\name{generate_hex}
\alias{generate_hex}
\title{Generates a default logo for a package.}
\usage{
generate_hex(
  name,
  output,
  colors = c("#1881C2", "#87B13F"),
  fallback_colors = c("#a60000", "#360000")
)
}
\arguments{
\item{name}{The package to generate the logo for.}

\item{output}{the path to store the generated logo at.}

\item{colors}{A vector of two valid colors. [1] for the fill and [2] for
the outline of the generated logo. Used if the package exists on CRAN.}

\item{fallback_colors}{A vector of two valid colors. [1] for the fill and
[2] for the outline of the generated logo. Used if the package
does not exist on CRAN.}
}
\value{
The path to the generated logo image.
}
\description{
Generates a generic logo for a package. IF the package is not on CRAN, a
  diferent set of colors is used.
}
\keyword{images}
\keyword{internal}
