% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather.R
\name{get_best_image}
\alias{get_best_image}
\title{Returns the best image path from a list of image paths}
\usage{
get_best_image(paths, download_endpoint, branch)
}
\arguments{
\item{paths}{A list of path information.}

\item{download_endpoint}{The download endpoint for the repo.}

\item{branch}{The repo active branch.}
}
\value{
A single url path
}
\description{
Returns the best image path from a list of image paths
}
\keyword{gather}
\keyword{internal}
