% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather.R
\name{get_cran_info}
\alias{get_cran_info}
\title{Returns CRAN information about a given package}
\usage{
get_cran_info(pkg_name)
}
\arguments{
\item{pkg_name}{A single string with the name of a package to find
the meta information for.}
}
\value{
A named list with CRAN metadata about the package
}
\description{
Returns CRAN information about a given package
}
\keyword{gather}
\keyword{internal}
