% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather.R
\name{get_endpoint_content}
\alias{get_endpoint_content}
\title{Fetch the content of a endpoint. Multiple branches can be given, the content
returned will also include wich was the valid branch.}
\usage{
get_endpoint_content(endpoint, branchs = c("master", "main"), token)
}
\arguments{
\item{endpoint}{The github endpoint to search}

\item{branchs}{The branches to search. Defaults to main and master.}

\item{token}{A github personal access token. Defaults to the enviromental}
}
\value{
A named list with the branch name and the repo content
}
\description{
Fetch the content of a endpoint. Multiple branches can be given, the content
returned will also include wich was the valid branch.
}
\keyword{gather}
\keyword{internal}
