% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather.R
\name{get_metacran_info}
\alias{get_metacran_info}
\title{Returns metacran information about a given package}
\usage{
get_metacran_info(pkg_name)
}
\arguments{
\item{pkg_name}{A single string with the name of a package to find
the meta information for.}
}
\value{
A named list with metadata information about the package
}
\description{
Returns metacran information about a given package
}
\keyword{gather}
\keyword{internal}
