% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather.R
\name{get_possible_paths}
\alias{get_possible_paths}
\title{Search the given content for paths that match the constrains from
  ignore_patterns and logo_patterns.}
\usage{
get_possible_paths(content, logo_patterns, ignore_patterns, pkg_name = "")
}
\arguments{
\item{content}{The result of calling get_endpoint_content() on a github repo.}

\item{logo_patterns}{A regex string of patterns to accept.}

\item{ignore_patterns}{A regex string of patterns to ignore.}

\item{pkg_name}{A Name of the package to find logos for.}
}
\value{
A list of paths that match the given patterns
}
\description{
Search the given content for paths that match the constrains from
  ignore_patterns and logo_patterns.
}
\keyword{gather}
\keyword{internal}
