% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather.R
\name{keep_good_ratio_images}
\alias{keep_good_ratio_images}
\title{Filters a list of image entries keeping images that might be a logo}
\usage{
keep_good_ratio_images(paths, download_endpoint, branch)
}
\arguments{
\item{paths}{A list of path information}

\item{download_endpoint}{The download endpoint for the repo}

\item{branch}{The repo active branch}
}
\value{
A list of path information
}
\description{
Filters a list of image entries keeping images that might be a logo
}
\keyword{gather}
\keyword{internal}
