% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locate.R
\name{locate_repo}
\alias{locate_repo}
\title{Finds the best repo for a given package}
\usage{
locate_repo(pkg_name)
}
\arguments{
\item{pkg_name}{A single string with the name of a package to find
the repo for.}
}
\value{
A URL that points to a github repository.
}
\description{
All CRAN packages have a cloned github repo that we can query for logos,
  but CRAN only stores the actual content of the package file. By looking
  at meta CRAN links we can try to find the original repo and use that one
  instead.
}
\keyword{internal}
\keyword{locator}
