% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log}
\alias{log}
\title{Logs a given message.}
\usage{
log(...)
}
\arguments{
\item{...}{The message to log.}
}
\value{
No return, called for side effects.
}
\description{
Logs a given message.
}
\keyword{internal}
\keyword{utils}
