% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scavenge.R
\name{scavenge}
\alias{scavenge}
\title{Same as [find_hex], but with extra Raccons}
\usage{
scavenge(...)
}
\arguments{
\item{...}{All arguments that can be passed to [find_hex].}
}
\value{
No return, called for side effects.
}
\description{
Same as [find_hex], but with extra Raccons
}
\keyword{external}
\keyword{finder}
