% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{url_file_exists}
\alias{url_file_exists}
\title{Checks if a given url is valid and available.}
\usage{
url_file_exists(url, status_ok = 200)
}
\arguments{
\item{url}{The resource url to check.}

\item{status_ok}{The list of valid header statuses.}
}
\value{
The result of evaluating the expression or NULL in case of error.
}
\description{
Checks if a given url is valid and available.
}
\keyword{internal}
\keyword{utils}
